+++ src/app/resourceView/resourceViewEditor/renderer/impl/evokedLexicalConceptsPartitionRenderer.ts
@@ -1,5 +1,7 @@
 import { Component } from "@angular/core";
 import { Observable, of } from "rxjs";
+import { OntoLexLemonServices } from "src/app/services/ontoLexLemonServices";
+import { NewConceptualizationCfModalReturnData } from "src/app/widget/modal/creationModal/newResourceModal/ontolex/newConceptualizationCfModal";
 import { ARTNode, ARTResource, ARTURIResource } from "../../../../models/ARTResources";
 import { ResViewPartition } from "../../../../models/ResourceView";
 import { CustomFormsServices } from "../../../../services/customFormsServices";
@@ -21,7 +23,8 @@ export class EvokedLexicalConceptsPartitionRenderer extends PartitionRenderSingl
     addBtnImgSrc = "./assets/images/icons/actions/objectProperty_create.png";
 
     constructor(resourcesService: ResourcesServices, propService: PropertyServices, cfService: CustomFormsServices,
-        basicModals: BasicModalServices, creationModals: CreationModalServices, resViewModals: ResViewModalServices,) {
+        basicModals: BasicModalServices, creationModals: CreationModalServices, resViewModals: ResViewModalServices,
+        private ontolexService: OntoLexLemonServices) {
         super(resourcesService, propService, cfService, basicModals, creationModals, resViewModals);
     }
 
@@ -30,7 +33,20 @@ export class EvokedLexicalConceptsPartitionRenderer extends PartitionRenderSingl
     }
 
     add(predicate: ARTURIResource, propChangeable: boolean) {
-        this.enrichProperty(predicate);
+        if (predicate.getURI() == this.rootProperty.getURI()) {
+            this.creationModals.newConceptualizationCf({key:"DATA.ACTIONS.ADD_CONCEPTUALIZATION"}, false, false).then(
+                (data: NewConceptualizationCfModalReturnData) => {
+                    this.ontolexService.addConceptualization(this.resource, data.linkedResource, true, data.createSense, data.cls, data.cfValue).subscribe(
+                        () => {
+                            this.update.emit()
+                        }
+                    );
+                },
+                () => {}
+            );
+        } else {
+            this.enrichProperty(predicate);
+        }
     }
 
     //not used since this partition doesn't allow manual add operation