+++ src/app/widget/modal/creationModal/newResourceModal/ontolex/newOntoLexicalizationCfModal.ts
@@ -17,25 +17,28 @@ export class NewOntoLexicalizationCfModal extends AbstractCustomConstructorModal
     @Input() lexicalizationProp: ARTURIResource; //(OntoLex.senses | OntoLex.denotes | Ontolex.isDenotedBy)
     @Input() clsChangeable: boolean = true;
 
+
+    /*
+    - ontolex:denotes => ask for Reference; checkbox: Create plain
+    - ontolex:isDenotedBy => ask for Lexical Entry; checkbox: Create plain
+    - ontolex:sense => ask for Reference OR Lexical Concept; checkbox: Create sense
+    */
+    showReference: boolean;
+    showLexEntry: boolean;
+    showRefAndConc: boolean; //case for ontolex:sense
+    showCreateSense: boolean;
+    showCreatePlain: boolean;
+
+    scenario: CreationScenario; //specify what is the resource to provide in the creation
+
     //standard form
     private linkedResource: string;
 
     private createPlainCheck: boolean = true;
     private createSenseCheck: boolean = true;
 
-    private pickerRoles: RDFResourceRolesEnum[] = [RDFResourceRolesEnum.cls, RDFResourceRolesEnum.individual, RDFResourceRolesEnum.property, 
+    pickerRoles: RDFResourceRolesEnum[] = [RDFResourceRolesEnum.cls, RDFResourceRolesEnum.individual, RDFResourceRolesEnum.property, 
         RDFResourceRolesEnum.concept, RDFResourceRolesEnum.conceptScheme, RDFResourceRolesEnum.skosCollection];
-
-    /**
-     * true if the modal should allow to link a lexical entry to a reference (show the reference input field),
-     * false if it should allow to link a resource to a lexical entry (show the lexical entry input field).
-     */
-    linkToReference: boolean;
-    /**
-     * true if the modal should allow to create a LexicalSense (show the "create plain" checkbox),
-     * false if it should allow to create a plain lexicalization (show the "create sense" checkbox).
-     */
-    createSense: boolean;
 
     constructor(public activeModal: NgbActiveModal, cfService: CustomFormsServices,
         basicModals: BasicModalServices, browsingModals: BrowsingModalServices) {
@@ -45,15 +48,18 @@ export class NewOntoLexicalizationCfModal extends AbstractCustomConstructorModal
     ngOnInit() {
         this.resourceClass = OntoLex.lexicalSense;
 
-        if (this.lexicalizationProp.getURI() == OntoLex.isDenotedBy.getURI()) {
-            this.linkToReference = false;
-            this.createSense = false;
+        if (this.lexicalizationProp.getURI() == OntoLex.denotes.getURI()) {
+            this.scenario = CreationScenario.reference;
+            this.showReference = true;
+            this.showCreatePlain = true;
+        } else if (this.lexicalizationProp.getURI() == OntoLex.isDenotedBy.getURI()) {
+            this.scenario = CreationScenario.lexEntry;
+            this.showLexEntry = true;
+            this.showCreatePlain = true;
         } else if (this.lexicalizationProp.getURI() == OntoLex.sense.getURI()) {
-            this.linkToReference = true;
-            this.createSense = true;
-        } else if (this.lexicalizationProp.getURI() == OntoLex.denotes.getURI()) {
-            this.linkToReference = true;
-            this.createSense = false;
+            this.scenario = CreationScenario.reference;
+            this.showRefAndConc = true;
+            this.showCreateSense = true;
         }
 
         this.selectCustomForm();
@@ -63,7 +69,7 @@ export class NewOntoLexicalizationCfModal extends AbstractCustomConstructorModal
         this.changeClassWithRoot(OntoLex.lexicalSense);
     }
 
-    private pickLexicalEntry() {
+    pickLexicalEntry() {
         this.browsingModals.browseLexicalEntryList({key:"DATA.ACTIONS.SELECT_LEXICAL_ENTRY"}).then(
             (lexEntry: ARTURIResource) => {
                 this.linkedResource = lexEntry.getURI();
@@ -71,7 +77,15 @@ export class NewOntoLexicalizationCfModal extends AbstractCustomConstructorModal
         )
     }
 
-    private updateLinkedRes(res: ARTURIResource) {
+    pickLexicalConcept() {
+        this.browsingModals.browseConceptTree({key:"DATA.ACTIONS.SELECT_LEXICAL_ENTRY"}).then(
+            (lexEntry: ARTURIResource) => {
+                this.linkedResource = lexEntry.getURI();
+            }
+        )
+    }
+
+    updateLinkedRes(res: ARTURIResource) {
         if (res != null) {
             this.linkedResource = res.getURI();
         } else {
@@ -84,9 +98,9 @@ export class NewOntoLexicalizationCfModal extends AbstractCustomConstructorModal
     }
 
     okImpl() {
-        var entryMap: any = this.collectCustomFormData();
+        let entryMap: any = this.collectCustomFormData();
 
-        var returnedData: NewOntoLexicalizationCfModalReturnData = {
+        let returnedData: NewOntoLexicalizationCfModalReturnData = {
             linkedResource: new ARTURIResource(this.linkedResource),
             createPlain: this.createPlainCheck,
             createSense: this.createSenseCheck,
@@ -98,7 +112,7 @@ export class NewOntoLexicalizationCfModal extends AbstractCustomConstructorModal
             returnedData.cls = this.resourceClass;
         }
         //set cfValue only if not null and only if it's creating a sense (that is a reified lexicalizaion, CF doesn't make sense for a plain lexicalization)
-        if (this.createSense && this.customFormId != null && entryMap != null) {
+        if (this.showCreateSense && this.customFormId != null && entryMap != null) {
             returnedData.cfValue = new CustomFormValue(this.customFormId, entryMap);
         }
         this.activeModal.close(returnedData);
@@ -116,4 +130,10 @@ export class NewOntoLexicalizationCfModalReturnData {
     createSense: boolean;
     cls: ARTURIResource
     cfValue: CustomFormValue;
+}
+
+enum CreationScenario {
+    reference = "reference", //ontolex:denotes and ontolex:sense
+    lexEntry = "lexEntry", //ontolex:isDenotedBy
+    lexConcept = "lexConcept" //ontolex:sense
 }