+++ ChangeLog.md
@@ -5,6 +5,7 @@ The full changelog of the platform includes also changes brought to the Semantic
 https://bitbucket.org/art-uniroma2/semantic-turkey/src/master/ChangeLog.txt
 
 # 10.0.1 (dd-mm-yyyy)
+  * Extended, in ResourceView, the creation of lexical sense allowing to use lexical concept.
   * Enabled multi-selection (ctrl+click) of languages in Project-User settings manager
   * Minor bugfixes and improvements